Int subtype double error fix + history page update

+++ lib/searchPage2.dart
@@ -85,7 +85,7 @@ class _HotelSearchPageState extends State<HotelSearchPage> {
     int numberOfNights = checkOutDate!.difference(checkInDate!).inDays;
     double baseRate = roomType == 'Traditional' ? pricePerNight : pricePerNight * 1.5;
     double additionalGuestCost = (numberOfGuests - 1) * 20.0;
-    double totalCost = (baseRate + additionalGuestCost) * numberOfNights;
+    double totalCost = (baseRate + additionalGuestCost) * numberOfNights.toDouble();
 
     return totalCost;
   }
